+++ src/main/java/com/braintreegateway/graphql/enums/InsightPaymentOption.java
@@ -1,0 +1,11 @@
+package com.braintreegateway.graphql.enums;
+
+/**
+ * Represents available payment options related to PayPal customer session insights.
+ *
+ * See our {@link https://graphql.braintreepayments.com/reference/#enum--insightpaymentoption graphql reference docs} for information on attributes.
+ */
+public enum InsightPaymentOption {
+  PAYPAL,
+  VENMO
+}